Add verification tests for 160M, 80M, and 60M delta loop presets
Investigated the reported rendering issue for lower HF band delta loop presets.
- Verified that `buildWires` generates valid geometry coordinates and segment counts for 160M, 80M, and 60M bands.
- Verified that the NEC-2 solver engine successfully simulates these presets.
- Added permanent verification tests to `tests/antennaStore.test.ts` to ensure ongoing validity.
- Removed temporary reproduction and investigation files.
- Verified all tests pass and linting is clean.

No application-level bugs were identified in the geometry or solver logic. The reported issue may be environment-specific.
