--- test-results/.last-run.json
@@ -1,6 +1,0 @@
-{
-  "status": "failed",
-  "failedTests": [
-    "bdbbf391b866701660d7-cee26313a797148570fe"
-  ]
-}
